add ref count

+++ middleware/chunk.js
@@ -12,6 +12,34 @@ const getChunkPath = function getChunkPath(chunkHash) {
     return chunkPath;
 };
 
+const getStatPath = function getStatPath(chunkHash) {
+    const chunkPath = path.join(config.upload.chunk, chunkHash.slice(0, 2), chunkHash.slice(2) + '.stat');
+    return chunkPath;
+};
+
+const getMetaPath = function getMetaPath(bucket, reqPath) {
+    const metaPath = path.join(bucket.path, reqPath);
+    return metaPath;
+};
+
+const readOrCreateStat = function readOrCreateStat(statPath) {
+    return co(function *() {
+        const stat = JSON.parse(yield fs.readFile(statPath, 'utf8'));
+        return stat;
+    }).catch((err) => {
+        if (err.errno === -2 && err.code === 'ENOENT') {
+            return [];
+        }
+        return err;
+    });
+};
+
+const saveStat = function saveStat(statPath, statContent) {
+    return co(function *() {
+        yield fs.writeFile(statPath, JSON.stringify(statContent), 'utf8');
+    });
+};
+
 const saveChunk = function saveChunk(req, res, next) {
     co(function *() {
         const singleFile = req.file;
@@ -20,6 +48,13 @@ const saveChunk = function saveChunk(req, res, next) {
 
         yield fs.mkdirp(path.dirname(newFilePath));
         yield file.stream(oldFilePath, newFilePath);
+
+        const statPath = getStatPath(singleFile.hash);
+        const stat = yield readOrCreateStat(statPath);
+        const metaPath = getMetaPath(res.body.bucket, req.path);
+        stat.push(metaPath);
+        yield saveStat(statPath, stat);
+
         next();
     }).catch((err) => {
         next(err);
@@ -38,6 +73,15 @@ const removeEmptyChunkFolder = function removeEmptyChunkFolder(chunkFolderPath) 
 
 const removeChunk = function removeChunk(req, res, next) {
     co(function *() {
+        const statPath = getStatPath(res.body.meta.hash);
+        let stat = yield readOrCreateStat(statPath);
+        const metaPath = getMetaPath(res.body.bucket, req.path);
+        stat = stat.filter((refPath) => refPath !== metaPath);
+        if (stat.length > 0) {
+            yield saveStat(statPath, stat);
+            return next();
+        }
+        yield fs.unlink(statPath);
         const meta = res.body.meta;
         const chunkPath = getChunkPath(meta.hash);
         yield fs.unlink(chunkPath);
@@ -50,6 +94,7 @@ const removeChunk = function removeChunk(req, res, next) {
 
 const sendChunk = function sendChunk(req, res, next) {
     const meta = res.body.meta;
+    if (!meta.hash) { return next(); }
     const chunkPath = getChunkPath(meta.hash);
     const options = {
         headers: {
